added options with unit tests

+++ index.js
@@ -1,17 +1,39 @@
 var rimraf = require('rimraf');
 var path = require('path');
 
-function Plugin(paths, context) {
+function Plugin(paths, options) {
+
+  //backwards compatibility
+  if (typeof options === 'string') {
+    options = {
+      root: options
+    }
+  }
+
+  options = options || {};
+  if (options.verbose === undefined) {
+    if (process.env.NODE_ENV === 'test') {
+      options.verbose = false;
+    } else {
+      options.verbose = true;
+    }
+  }
+
+  if (options.dry === undefined) {
+    options.dry = false;
+  }
+
   // determine webpack root
-  this.context = context || path.dirname(module.parent.filename);
+  options.root = options.root || path.dirname(module.parent.filename);
 
   // allows for a single string entry
   if (typeof paths == 'string' || paths instanceof String) {
     paths = [paths];
   }
 
-  // store paths
+  // store paths and options
   this.paths = paths;
+  this.options = options;
 }
 
 Plugin.prototype.apply = function(compiler) {
@@ -25,24 +47,26 @@ Plugin.prototype.apply = function(compiler) {
     return 'nothing to clean';
   }
 
-  if (!path.isAbsolute(self.context)) {
-    console.warn('clean-webpack-plugin: ' + self.context + ' project root must be an absolute path. Skipping all...');
+  if (!path.isAbsolute(self.options.root)) {
+    self.options.verbose && console.warn(
+      'clean-webpack-plugin: ' + self.options.root + ' project root must be an absolute path. Skipping all...');
     return 'project root must be an absolute path';
   }
 
-  var projectDir = path.resolve(self.context);
+  var projectDir = path.resolve(self.options.root);
   var projectDirSplit = projectDir.split('/');
   var webpackDir = path.dirname(module.parent.filename);
 
   // This is not perfect.
   if ((projectDir !== __dirname.slice(0, projectDir.length)) && (projectDir.indexOf(__dirname) === -1)) {
-    console.warn('clean-webpack-plugin: ' + self.context + ' project root is outside of project. Skipping all...');
+    self.options.verbose &&
+    console.warn('clean-webpack-plugin: ' + self.options.root + ' project root is outside of project. Skipping all...');
     return 'project root is outside of project';
   }
 
   // preform an rm -rf on each path
   self.paths.forEach(function(rimrafPath) {
-    rimrafPath = path.resolve(self.context, rimrafPath);
+    rimrafPath = path.resolve(self.options.root, rimrafPath);
 
     pathSplit = rimrafPath.split('/');
     projectDirSplit.forEach((function(singleDir, index) {
@@ -52,6 +76,7 @@ Plugin.prototype.apply = function(compiler) {
     }));
 
     if (insideFailCheck) {
+      self.options.verbose &&
       console.warn('clean-webpack-plugin: ' + rimrafPath + ' must be inside the project root. Skipping...');
       results.push({path: rimrafPath, output: 'must be inside the project root'});
       return;
@@ -60,12 +85,14 @@ Plugin.prototype.apply = function(compiler) {
     // disallow deletion of project path and any directories outside of project path.
     if (rimrafPath.indexOf(projectDir) === 0) {
       if (rimrafPath === projectDir) {
+        self.options.verbose &&
         console.warn('clean-webpack-plugin: ' + rimrafPath + ' is equal to project root. Skipping...');
         results.push({path: rimrafPath, output: 'is equal to project root'});
         return;
       }
 
       if (rimrafPath === webpackDir) {
+        self.options.verbose &&
         console.warn('clean-webpack-plugin: ' + rimrafPath + ' would delete webpack. Skipping...');
         results.push({path: rimrafPath, output: 'would delete webpack'});
         return;
@@ -73,12 +100,17 @@ Plugin.prototype.apply = function(compiler) {
 
       // If path is completely outside of the project. Should not be able to happen. Including as an extra fail-safe.
       if (projectDir.indexOf(rimrafPath) !== -1) {
+        self.options.verbose &&
         console.warn('clean-webpack-plugin: ' + rimrafPath + ' is unsafe to delete. Skipping...');
         results.push({path: rimrafPath, output: 'is unsafe to delete'});
         return;
       }
 
-      rimraf.sync(rimrafPath);
+      if (self.options.dry !== true) {
+        rimraf.sync(rimrafPath);
+      }
+
+      self.options.verbose && console.log('clean-webpack-plugin: ' + rimrafPath + ' has been removed.');
       results.push({path: rimrafPath, output: 'removed'});
     }
   });
